requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
discord.py==1.4.1
dispander==0.4.0
echidna==0.2.1
flake8==3.8.3
# -e git+https://github.com/Hukasan/discord-bot-id.git@478b80a410276f3d0f306a804a24f91f73e40363#egg=Flask
gunicorn==20.0.4
Hamlish-Jinja==0.3.3
httplib2==0.18.1
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
pipdeptree==1.0.0
# pkg-resources==0.0.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.5.3
PyNaCl==1.4.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
typing-extensions==3.7.4.2
tzlocal==2.1
urllib3==1.25.9
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.0
emoji==0.6.0
